+++ roles/keepalived/filter_plugins/manala_environment.py
@@ -27,11 +27,13 @@ def environment_parameters(parameters, exclude=[]):
     return result
 
 
-def environment_parameter(parameters, key, default=None, comment=False):
+def environment_parameter(parameters, key, required=False, default=None, comment=False):
     if not isinstance(parameters, dict):
         raise AnsibleFilterError('manala_environment_parameter parameters expects a dict but was given a %s' % type(parameters))
     if not isinstance(key, string_types):
         raise AnsibleFilterError('manala_environment_parameter key expects a string but was given a %s' % type(key))
+    if required and key not in parameters:
+        raise AnsibleFilterError('manala_environment_parameter requires a value for key %s' % key)
     result = ''
     value = parameters.get(key, default)
     if isinstance(value, string_types):