+++ roles/keepalived/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Handle required parameters keys in environment filters
 
 ## [1.0.7] - 2020-12-02
 ### Added