+++ roles/rsyslog/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Handle required parameters keys in config filters
 
 ## [2.0.4] - 2020-10-28
 ### Changed